+++ spec/mongoid/relations/referenced/many_to_many_spec.rb
@@ -18,6 +18,40 @@ describe Mongoid::Relations::Referenced::ManyToMany do
 
     describe "##{method}" do
 
+      context "when the inverse_of is nil" do
+
+        let!(:article) do
+          Article.create
+        end
+
+        context "when the child document is new" do
+
+          let(:preference) do
+            Preference.new
+          end
+
+          before do
+            article.preferences.send(method, preference)
+          end
+
+          it "persists the child document" do
+            preference.should be_persisted
+          end
+        end
+
+        context "when the child document is not new" do
+
+          let(:preference) do
+            Preference.create
+          end
+
+          it "does not persist the child document" do
+            preference.should_receive(:save).never
+            article.preferences.send(method, preference)
+          end
+        end
+      end
+
       context "when the parent is a new record" do
 
         let(:person) do